add static_cast

+++ src/types/redis_list.cc
@@ -335,7 +335,7 @@ rocksdb::Status List::Insert(const Slice &user_key, const Slice &pivot, const Sl
   metadata.Encode(&bytes);
   batch->Put(metadata_cf_handle_, ns_key, bytes);
 
-  *new_size = metadata.size;
+  *new_size = static_cast<int>(metadata.size);
   return storage_->Write(storage_->DefaultWriteOptions(), batch->GetWriteBatch());
 }
